# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Описание работы сервера",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/decoration/delete": {
            "delete": {
                "consumes": [
                    "application/text"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cake Decorations"
                ],
                "summary": "Delete a given cake decoration by its article",
                "parameters": [
                    {
                        "description": "Article of a cake decoration to be deleted",
                        "name": "Article",
                        "in": "body",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/decoration/edit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cake Decorations"
                ],
                "summary": "Edit a given decoration",
                "parameters": [
                    {
                        "description": "scheme of cake decoration",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CakeDecoration"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.CakeDecoration"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/decoration/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cake Decorations"
                ],
                "summary": "Get a list of all cake decorations",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.CakeDecoration"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/failure/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Failures"
                ],
                "summary": "List all failures occurred in the system",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Failure"
                            }
                        }
                    },
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/failure/register": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Failures"
                ],
                "summary": "Register a new failure manually",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Failure"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/order/history": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get all orders history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.OrdersHistory"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CakeDecoration": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "article": {
                    "type": "string"
                },
                "costPrice": {
                    "type": "number"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "supplier": {
                    "$ref": "#/definitions/domain.Supplier"
                },
                "supplierName": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "unit": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Failure": {
            "type": "object",
            "properties": {
                "continueTime": {
                    "type": "string"
                },
                "equipment": {
                    "type": "string"
                },
                "failureTime": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "domain.Item": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "size": {
                    "type": "string"
                }
            }
        },
...
        "domain.Order": {
            "type": "object",
            "properties": {
                "assignedManager": {
                    "$ref": "#/definitions/domain.User"
                },
                "assignedManagerName": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "examples": {
                    "type": "string"
                },
                "expectedFulfilmentDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "item": {
                    "$ref": "#/definitions/domain.Item"
                },
                "itemName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orderer": {
                    "$ref": "#/definitions/domain.User"
                },
                "ordererName": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "domain.OrdersHistory": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "newStatus": {
                    "type": "string"
                },
                "oldStatus": {
                    "type": "string"
                },
                "order": {
                    "$ref": "#/definitions/domain.Order"
                },
                "orderId": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.Supplier": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "deliveryTime": {
                    "description": "в часах",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "domain.User": {
            "type": "object",
            "properties": {
                "fullName": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "photoURL": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        }
    },
...
}